# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Keues API",
    "description": "Queue and ticket management API: locations, flows, queues, counters, tickets and devices. Authentication is done via the HttpOnly \"access_token\" cookie (JWT) set by the login and create-admin endpoints.",
    "contact": {
      "name": "Keues",
      "url": "https://www.keues.dev"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5125",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/Counters": {
      "post": {
        "tags": [
          "Counters"
        ],
        "summary": "Creates a new service desk.",
        "requestBody": {
          "description": "Data of the counter to create.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCounterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCounterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCounterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Counter created.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Counters"
        ],
        "summary": "Gets all counters desks, optionally filtered by location.",
        "parameters": [
          {
            "name": "LocationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of counters.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfCounterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfCounterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseOfIEnumerableOfCounterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Counters/{id}": {
      "put": {
        "tags": [
          "Counters"
        ],
        "summary": "Updates an existing service desk.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the counter.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Data to update the counter.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCounterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCounterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCounterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Counter updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Counters"
        ],
        "summary": "Deletes a counter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the counter.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Counter deleted."
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          },
          {
            "access_token": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Counters"
        ],
        "summary": "Gets a counter by its identifier.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the counter.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Counter found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Counters/{id}/call-next-ticket": {
      "post": {
        "tags": [
          "Counters"
        ],
        "summary": "Calls the next ticket in the queue for the specified counter and notifies\nthe group's monitors (SignalR, \"TicketCalled\" event).",
        "description": "If the counter already has a ticket in progress, the same ticket is called again.\nIf there are no tickets waiting for the counter's queues, returns 200 with null body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the counter making the call.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "FlowId used to build the SignalR group.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallNextTicketRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CallNextTicketRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CallNextTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ticket called (may be null).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CallNextTicketResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallNextTicketResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallNextTicketResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Counters/{id}/attend-ticket": {
      "post": {
        "tags": [
          "Counters"
        ],
        "summary": "Marks a ticket as attended and notifies the group's monitors\n(SignalR, \"TicketAttended\" event).",
        "description": "The counter is resolved from the path id (for the SignalR group),\nwhile the command uses the CounterId from the body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the counter that attends.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "CounterId, TicketId and FlowId of the operation.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendTicketRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendTicketRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AttendTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ticket attended."
          },
          "400": {
            "description": "Validation or business rule error.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCounterRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "queues": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Colas a las que tiene acceso este counter"
          },
          "authorizedUsers": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Usuarios autorizados para este counter"
          },
          "authorizedUserGroups": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Grupos de usuarios autorizados para este counter"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "access_token": {
        "type": "apiKey",
        "description": "HttpOnly cookie with the JWT. It is obtained by calling POST /api/users/login or POST /api/users/create-admin.",
        "name": "access_token",
        "in": "cookie"
      },
      "bearer": {
        "type": "http",
        "description": "JWT sent in the \"Authorization: Bearer <token>\" header. Same token returned in the body by POST /api/users/login and POST /api/users/create-admin; use it for non-browser clients that cannot rely on the HttpOnly cookie.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}